Adding account template from keycloak.v2
adding account in simva theme from keycloak theme

update theme parent

fix theme account properties

adding index for keycloak account page

comeback to keycloak v2 theme

adding all from keycloak v2

fixing some items for account page

set to development mode

fixing script in case folder don't exist

+++ docker-stacks/02-keycloak/etc/hooks/purge.d/02-remove-extensions-target.sh
@@ -1,6 +1,8 @@
 # Removing Extension data
 EXTENSIONS_DIR="${STACK_HOME}/extensions"
 for extension in $(find ${EXTENSIONS_DIR} -mindepth 1 -maxdepth 1 -type d); do
-    cd "${extension}/target"
-    rm -rf ./*
+    if [[ -e "${extension}/target" ]]; then
+        cd "${extension}/target"
+        rm -rf ./*
+    fi
 done
